# backend/docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "PRM Marketplace API",
    "description" : "Backend API for the student/resident marketplace",
    "contact" : {
      "name" : "PRM Marketplace Team"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/products/{id}" : {
      "get" : {
        "tags" : [ "product-controller" ],
        "operationId" : "read_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "product-controller" ],
        "operationId" : "update_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Product"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/api/product-images/{id}" : {
      "get" : {
        "tags" : [ "product-image-controller" ],
        "operationId" : "read_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductImage"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "product-image-controller" ],
        "operationId" : "update_4",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProductImage"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductImage"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "product-image-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/products" : {
      "get" : {
        "tags" : [ "product-controller" ],
        "operationId" : "getAll_4",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "product-controller" ],
        "operationId" : "create_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Product"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/api/product-images/product/{productId}" : {
      "get" : {
        "tags" : [ "product-image-controller" ],
        "operationId" : "getByProduct",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProductImage"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "product-image-controller" ],
        "operationId" : "create_5",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProductImage"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductImage"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "product-image-controller" ],
        "operationId" : "deleteByProduct",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/reviews/product/{productId}" : {
      "get" : {
        "tags" : [ "review-controller" ],
        "operationId" : "getByProduct_1",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products/vendor/{vendorId}" : {
      "get" : {
        "tags" : [ "product-controller" ],
        "operationId" : "getByVendor",
        "parameters" : [ {
          "name" : "vendorId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/search" : {
      "get" : {
        "tags" : [ "product-controller" ],
        "operationId" : "search",
        "parameters" : [ {
          "name" : "keyword",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "category",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "city",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "minPrice",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "number"
          }
        }, {
          "name" : "maxPrice",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "number"
          }
        }, {
          "name" : "condition",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "NEW", "LIKE_NEW", "GOOD", "FAIR", "POOR" ]
          }
        }, {
          "name" : "activeOnly",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "createdAt"
          }
        }, {
          "name" : "direction",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "desc"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseProduct"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products/category/{category}" : {
      "get" : {
        "tags" : [ "product-controller" ],
        "operationId" : "getByCategory",
        "parameters" : [ {
          "name" : "category",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/product-images/product/{productId}/primary" : {
      "get" : {
        "tags" : [ "product-image-controller" ],
        "operationId" : "getPrimary",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductImage"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/saved-items/product/{productId}" : {
      "delete" : {
        "tags" : [ "saved-item-controller" ],
        "operationId" : "removeByUserAndProduct",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "VendorProfile" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "readOnly" : true
          },
          "businessName" : {
            "type" : "string"
          },
          "registrationNo" : {
            "type" : "string"
          },
          "verified" : {
            "type" : "boolean",
            "readOnly" : true
          },
          "ratingAvg" : {
            "type" : "number",
            "readOnly" : true
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          }
        }
      },
      "Review" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "productId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reviewerId" : {
            "type" : "string",
            "format" : "uuid",
            "readOnly" : true
          },
          "rating" : {
            "type" : "integer",
            "format" : "int32"
          },
          "comment" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          }
        }
      },
      "Product" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number"
          },
          "stockQuantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "category" : {
            "type" : "string"
          },
          "imageUrl" : {
            "type" : "string"
          },
          "condition" : {
            "type" : "string",
            "enum" : [ "NEW", "LIKE_NEW", "GOOD", "FAIR", "POOR" ]
          },
          "city" : {
            "type" : "string"
          },
          "province" : {
            "type" : "string"
          },
          "vendor" : {
            "$ref" : "#/components/schemas/VendorProfile",
            "readOnly" : true
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          },
          "active" : {
            "type" : "boolean"
          }
        }
      },
      "ProductImage" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "product" : {
            "$ref" : "#/components/schemas/Product",
            "writeOnly" : true
          },
          "imageUrl" : {
            "type" : "string"
          },
          "sortOrder" : {
            "type" : "integer",
            "format" : "int32"
          },
          "primary" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          }
        }
      },
...
      "PageResponseProduct" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Product"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}